JAMES-1759 rename model package into dto
