+++ server/protocols/webadmin/src/main/java/org/apache/james/webadmin/dto/AddUserRequest.java
@@ -17,17 +17,23 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.webadmin.model;
+package org.apache.james.webadmin.dto;
 
-public class MailboxResponse {
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.google.common.base.Preconditions;
 
-    private final String mailboxName;
+public class AddUserRequest {
 
-    public MailboxResponse(String mailboxName) {
-        this.mailboxName = mailboxName;
+    private final char[] password;
+
+    @JsonCreator
+    public AddUserRequest(@JsonProperty("password") char[] password) {
+        Preconditions.checkNotNull(password);
+        this.password = password;
     }
 
-    public String getMailboxName() {
-        return mailboxName;
+    public char[] getPassword() {
+        return password;
     }
 }
